+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 0.21.1
+
 ## 0.21.0 (October 14, 2013)
 
 ### Enhancements